+++ sdk/storage/azure-storage-blob/tests/test_blob_samples_service_async.py
@@ -51,10 +51,10 @@ class TestBlobServiceSamplesAsync(StorageTestCase):
 
         # [START set_blob_service_properties]
         # Create service properties
-        from azure.storage.blob import Logging, Metrics, CorsRule, RetentionPolicy
+        from azure.storage.blob import BlobAnalyticsLogging, Metrics, CorsRule, RetentionPolicy
 
         # Create logging settings
-        logging = Logging(read=True, write=True, delete=True, retention_policy=RetentionPolicy(enabled=True, days=5))
+        logging = BlobAnalyticsLogging(read=True, write=True, delete=True, retention_policy=RetentionPolicy(enabled=True, days=5))
 
         # Create metrics for requests statistics
         hour_metrics = Metrics(enabled=True, include_apis=True, retention_policy=RetentionPolicy(enabled=True, days=5))